Create main-process/settings/restore/window.ts
Export async function as default
